chore(supabase): give Libra its own local instance, independent of any other [GH-000]
`pnpm supabase:start` was starting the CLI's default instance on base port
54321, not this project's. Two consequences: it collided with any other
Supabase on the machine (AeleOS runs there), and when it did come up it was a
different database from the one .env.dev pointed at.

The cause is that supabase/config.toml is generated from
config.toml.template and gitignored, but only supabase-docker.mjs generated
it. supabase-cmd.mjs -- which backs supabase:start, :stop and :reset -- ran
the CLI with no config at all, so the CLI used its own defaults.

The generator is now scripts/lib/supabase-config.mjs, shared by both entry
points rather than copied. supabase:start generates it and deliberately leaves
it in place, so direct CLI calls that bypass the wrapper -- `supabase gen
types --local` in codegen:supabase, or `supabase status` by hand -- also target
this project's instance instead of silently reading the wrong one.

Dev now points at that instance: http://localhost:54331, with the standard
Supabase local keys (issued by "supabase-demo", signed with the demo JWT
secret already in this file, and published in Supabase's own docs -- not
secrets). SUPABASE_MODE moves from cloud to docker to match.

That also fixes a live breakage: .env.dev pointed at the DEV_SUPABASE_* cloud
project, which no longer exists (docs/production-status.md). Anyone running
`pnpm dev` was talking to a dead host. The secrets are left in .secrets but
are no longer read.

Ports derive from SUPABASE_PORT=54331, which was already in .env.dev and
already matched the running containers -- api 54331, db 54332, studio 54333,
inbucket 54334, analytics 54337. CI is untouched: .env.ci keeps 54321, and
generating its config still reports project libra-ci on 54321.

Verified: both instances answer simultaneously (54321 AeleOS, 54331 Libra,
12 containers each); user_profiles, products and permissions all reachable
with the dev env loaded; lint:env, typecheck, lint, format:check.
