+++ .env.dev
@@ -5,7 +5,7 @@ TARGET_ENV=dev
 
 # --- Topology -----------------------------------------------------------------
 APPS_MODE=local
-SUPABASE_MODE=cloud
+SUPABASE_MODE=docker
 
 # --- Container identity (unused in dev, kept for parity) ----------------------
 SITE_PROD_CONTAINER_NAME=libra-dev
@@ -15,13 +15,22 @@ SITE_PROD_IMAGE_NAME=libra-dev
 HOST_PORT=5050
 APP_INTERNAL_ORIGIN=http://localhost:8088
 
-# --- Supabase Cloud (dev project) ---------------------------------------------
-NEXT_PUBLIC_SUPABASE_URL=$secret:DEV_SUPABASE_URL
-SUPABASE_URL_INTERNAL=
-NEXT_PUBLIC_SUPABASE_ANON_KEY=$secret:DEV_SUPABASE_ANON_KEY
-SUPABASE_SERVICE_ROLE_KEY=$secret:DEV_SUPABASE_SERVICE_ROLE_KEY
+# --- Supabase (local, this project only) ---------------------------------------
+# Runs on its own base port so it never collides with another project's local
+# Supabase. `pnpm supabase:start` generates supabase/config.toml from
+# SUPABASE_PORT, so every service lands on 5433x rather than the CLI default
+# 5432x. Started as project "libra-dev"; `docker ps` shows *_libra-dev.
+#
+# The cloud dev project these used to point at (DEV_SUPABASE_* in .secrets) no
+# longer exists -- see docs/production-status.md. Those secrets are left in
+# place but are no longer read here.
+NEXT_PUBLIC_SUPABASE_URL=http://localhost:54331
+SUPABASE_URL_INTERNAL=http://host.docker.internal:54331
+# Standard Supabase local keys: signed with the demo JWT secret below, issued by
+# "supabase-demo", and published in Supabase's own docs. Not secrets.
+NEXT_PUBLIC_SUPABASE_ANON_KEY=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZS1kZW1vIiwicm9sZSI6ImFub24iLCJleHAiOjE5ODM4MTI5OTZ9.CRXP1A7WOeoJeXxjNni43kdQwgnWNReilDMblYTn_I0
+SUPABASE_SERVICE_ROLE_KEY=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZS1kZW1vIiwicm9sZSI6InNlcnZpY2Vfcm9sZSIsImV4cCI6MTk4MzgxMjk5Nn0.EGIM96RAZx35lJzdJsyH-qQwv8Hdp7fsn3W0YpN81IU
 
-# Not used in cloud mode. Kept for env parity with non-cloud environments.
 SUPABASE_PORT=54331
 SUPABASE_JWT_SECRET=super-secret-jwt-token-with-at-least-32-characters-long
 SUPABASE_DB_PASSWORD=postgres
